FIX! (cgi) avoid child killing process of other request and first message
Added a boolean 'isChildren'. So that destructor of child process can
avoid doing actions like killing process of other requests.
Destructor of 'Connection' do not remove process from interest list of
epoll anymore, since children would remove fd concerning In/OutCGI
which would break the program
First part of body read with the headers of a cgi is now happended to
the first buffer being send back to the client (in case everything is in
the first buffer).

+++ sources/classes/Connection/Connection.cpp
@@ -49,7 +49,7 @@ Connection::~Connection()
 	//this->server.remove(&inputSocket);
 	//this->server.remove(&outputSocket);
 
-	if (this->cgiPid > 0)
+	if (this->server.getIsChildren() == false && this->cgiPid > 0)
 	{
 		if (kill(this->cgiPid, SIGTERM) < 0)
 		{
@@ -59,14 +59,18 @@ Connection::~Connection()
 		this->cgiPid = -1;
 	}
 	if (this->inCGI != NULL)
-		this->server.remove(this->inCGI);
-	delete this->inCGI;
-	this->inCGI = NULL;
+	{
+		//this->server.remove(this->inCGI);
+		delete this->inCGI;
+		this->inCGI = NULL;
+	}
 
 	if (this->outCGI != NULL)
-		this->server.remove(this->outCGI);
-	delete this->outCGI;
-	this->outCGI = NULL;
+	{
+		//this->server.remove(this->outCGI);
+		delete this->outCGI;
+		this->outCGI = NULL;
+	}
 }
 
 void	Connection::add(ASocket* abstractSocket, int event)
@@ -76,7 +80,7 @@ void	Connection::add(ASocket* abstractSocket, int event)
 
 void	Connection::remove(ASocket* abstractSocket)
 {
-	if (this->cgiPid > 0)
+	if (this->server.getIsChildren() == false && this->cgiPid > 0)
 	{
 		if (kill(this->cgiPid, SIGTERM) < 0)
 		{
@@ -85,6 +89,11 @@ void	Connection::remove(ASocket* abstractSocket)
 		}
 	}
 	this->server.remove(abstractSocket);
+}
+
+void			Connection::setIsChildren()
+{
+	this->server.setIsChildren();
 }
 
 InputSocket*	Connection::getInputSocket() {return (&this->inputSocket);}
